Add Edge Computing as a new category and StackPath Propel Program
Hi there, thanks for putting this together it's such a great resource. I'd like to propose adding Edge Compute as a new category and add the Stackpath Propel program to that category. The propel program provides startups with up to $2000 of platform consumption for 12 months and gives the company access to 24/7 support and mentorship from seasoned entrepreneurs and engineers. 
Full disclosure, I work for StackPath and help manage the program.